+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v0.9.41
+- Add revert if Seaport order is executed outside `SeaportProxyBuyer` contract
 - Added a script to batch mint TLDs proposal
 
 ## v0.9.40